example-app: twister-verifiable tests of subsys custom_subsys
This change provides CI-discoverable and verifiable content in the
`include/`, `subsys/`, and `tests/` directories of the Zephyr tree.
`testcase.yaml` files are discoverable and runnable by
`twister -T example-application`. The tests themselves verify the
`include/` tree is visble through the build system and that the
subsystem is buildable and runnable.  `module.yml` has been configured
so the `testcase.yaml` should be automatically discoverable without
explicit `twister` parameters.
.
This approach was chosen to provide examples in Zephyr directory
sub-trees not already provided, to provide twister-runnable tests
(not already provided), and to provide valid API documentation files
as a test case for Zephyr's document generation tools.
.
It appears no assumptions were made in this design.
.
The implementation was verified in the following ways:
  - Ran
        `./zephyr/scripts/twister -v \
         --tag extensibility --tag introduction \
         -T example-application/ \
         --dry-run`
    - Verified that `twister-out/twister.log` found possible
      test cases under `example-application/tests/`, and did not
      discard all of them due to filters.
  - Ran the above command without `--dry-run`.
    - Verified that it reported PASSED on >=1 non-build-only `tests/` job.
      (It passed 18/21, including running on qemu_cortex_m0.)
  - Ran
        `   cp -r example-application/include/custom_subsys \
                  zephyr/include/ \
         && cd zephyr/doc && make clean && make && cd ../.. \
         && rm -rf zephyr/include/custom_subsys \
        `
    - Verified "custom_subsys_said" was found in the generated documentation,
      confirming the header was valid if it were in the Zephyr repo.

Fixes #33962

Signed-off-by: Gregory Shue <gregory.shue@legrand.us>