example-app: twister-runnable sample app_empty_main
This change provides CI-discoverable and verifiable content in
the `samples/` directory of the Zephyr tree.  It provides a
`sample.yaml` file that is discoverable and runnable by
`twister -T example-application`.  `module.yml` has been configured
so the `sample.yaml` should be automatically discoverable without
explicit `twister` parameters.
.
This approach was chosen to provide examples in Zephyr directory
sub-trees not already provided, to provide twister-runnable sample(s)
(not already provided), and to provide valid subsys documentation
files as a test case for Zephyr's document generation tools.
.
It appears no assumptions were made in this design.
.
The implementation was verified in the following ways:
  - Ran
        `./zephyr/scripts/twister -v \
         --tag extensibility --tag introduction \
         -T example-application/samples/ \
         --dry-run`
    - Verified that `twister-out/twister.log` found possible
      test cases under `example-application/samples/`, and
      did not discard all of them due to filters.
  - Ran the above command without `--dry-run`.
    - Verified that it reported PASSED on >=1 build-only `samples/` job
      (PASSED 5/5).
    - Verified that it reported PASSED or FAILED Timeout on >=1
      non-build-only `samples/` job  (FAILED Timeout on 13/16).
  - Ran
        `   mkdir zephyr/samples/modules/example-application \
         && cp -r example-application/samples/* \
                  zephyr/samples/modules/example-application/ \
         && cd zephyr/doc && make clean && make && cd ../.. \
         && rm -rf zephyr/samples/modules/example-application \
        `
    - Verified the "App Empty Main" sample was listed in the
      Samples and Demos > External Module Samples > App Empty Main
      section of the generated document, confirming the README.txt
      is valid.

Signed-off-by: Gregory Shue <gregory.shue@legrand.us>